@@ -2,10 +2,8 @@ package org.nucleodevel.cointrader.beans;
 
 public enum Provider {
 
-	MERCADO_BITCOIN("MERCADO_BITCOIN"),
-	BLINKTRADE("BLINKTRADE"),
-	POLONIEX("POLONIEX");
-	
+	MERCADO_BITCOIN("MERCADO_BITCOIN"), BLINKTRADE("BLINKTRADE"), POLONIEX("POLONIEX");
+
 	private final String value;
 
 	private Provider(String value) {
